smartdine/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
sentence-transformers
faiss-cpu
python-dotenv
requests
groq
mysql-connector-python
